refactor: keyGenerator class를 util 에서 component로 이동
